Reorganized the file hierarchy and added in ReactRouter along with deleting some unneeded files.  No real possibilities for bugs with this commit(dangerous statement, I know)
