+++ src/index.js
@@ -1,11 +1,16 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
-import * as serviceWorker from './serviceWorker';
+import Routes from './Routes';
+import {BrowserRouter} from 'react-router-dom';
+import ApolloClient from "apollo-boost";
+import {ApolloProvider} from "react-apollo";
 
 
-ReactDOM.render(<App/>, document.getElementById('root'));
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: http://bit.ly/CRA-PWA
-serviceWorker.unregister();
+const client = new ApolloClient({uri: "http://localhost:8000/api/graph"});
+
+ReactDOM.render(<BrowserRouter>
+    <ApolloProvider client={client}>
+        <Routes/>
+    </ApolloProvider>
+</BrowserRouter>, document.getElementById('root'));
+
