+++ src/pages/BasicTextPage.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export const Text = () => {
+    return (
+        <div>
+            <p>TextPage</p>
+
+        </div>
+    )
+};
+
+export default Text;
