+++ src/pages/OrganizationCatalog.js
@@ -1,13 +1,11 @@
-import React, {Component} from 'react';
-import {ApolloProvider} from "react-apollo";
-import ApolloClient from "apollo-boost";
-import getAllOrganizations from "./getAllOrganizations.js";
+import React from 'react';
+import getAllOrganizations from "../queries/getAllOrganizations.js";
 import {Item} from 'semantic-ui-react'
 import 'semantic-ui-css/semantic.min.css'
 import {isNullOrUndefined} from "util";
 
-const client = new ApolloClient({uri: "http://localhost:8000/api/graph"});
-let default_image = require("./defaultuserimg.jpg");
+
+let default_image = require("../resources/defaultImage.jpg");
 const orgCard = (agent) => (
     <Item>
         <Item.Image src={isNullOrUndefined(agent.image) || agent.image === "" ? default_image : agent.image}/>
@@ -50,15 +48,18 @@ const OrgList = getAllOrganizations(({organizationList, loading, error}) => {
     }
 });
 
-class App extends Component {
+class BasePage extends React.Component {
     render() {
         return (
-            <ApolloProvider client={client}>
+            <div>
                 <OrgList
                     token="eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImNvbm5vciIsImlhdCI6MTU1MTg0ODI3MSwicGFzc3dvcmQiOiI3YzA4ODliOWU5ZmNjYzAxZDIzMDcwNzljNDk5OTcyNDFlNTZlNzU0IiwiaWQiOjZ9.unIuk6g8HcmyIuF1sONrLAiftApTlcuqMWWLO6DtqUQ"/>
-            </ApolloProvider>
-        );
+            </div>
+        )
     }
 }
 
-export default App;
+
+
+
+export default BasePage;
