+++ src/resources/defaultImage.jpg
